api/serializers.py
# from rest_framework import serializers
from data.models import Userdata
from rest_framework import serializers

class UserSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    class Meta:
        model = Userdata
        fields = ('name', 'username', 'email', 'phone','college','junior', 'password')

    def create(self, validated_data):
        user = super(UserSerializer, self).create(validated_data)
        user.set_password(validated_data['password'])
        user.save()
        return user