requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
Cython==0.29.21
distlib==0.3.0
distro==1.4.0
Django==3.1.4
djangorestframework==3.12.4
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jdcal==1.4.1
lockfile==0.12.2
msgpack==0.6.2
mysqlclient==2.0.3
openpyexcel==2.5.14
openpyxl==3.0.6
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
seccomp==3.8
six==1.14.0
sqlparse==0.4.1
tz==0.2.2
urllib3==1.25.8
webencodings==0.5.1
